Add process offset as well.
